+++ src/core/Mida.ts
@@ -43,7 +43,7 @@ class Mida {
     static readonly #installedPlugins: Map<string, MidaPlugin> = new Map();
 
     public static get version (): string {
-        return "2022.2.0";
+        return "2022.3.0";
     }
 
     public static get logger (): MidaLogger {