+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Features
+* Expose the queue API [#66](https://github.com/Reiryoku-Technologies/Mida/pull/66)
+
 2022.2.0 - 04-08-2022
 ===================
 ### Features