+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+2022.3.0 - 12-08-2022
+===================
 ### Features
 * Add native `period-close` event usage to market watcher, allow to optionally use the builtin period close detector [#67](https://github.com/Reiryoku-Technologies/Mida/pull/67)
 * Expose the queue API [#66](https://github.com/Reiryoku-Technologies/Mida/pull/66)